+++ pkgs/test/CHANGELOG.md
@@ -13,6 +13,8 @@
 * Update all reporters to print the compiler along with the platform name
   when configured to print the platform. Extend the logic for printing platofrm
   information to do so if any compilers are explicitly configured.
+* Deprecate `--use-data-isolate-strategy`. It is now an alias for `-c vm:source`
+  which is roughly equivalent. If this is breaking for you please file an issue.
 
 ## 1.23.1